+++ R/learndown-package.R
@@ -9,9 +9,9 @@
 #' @importFrom glue glue
 #' @importFrom webshot webshot
 #' @importFrom magick image_read image_write image_composite image_info image_crop
-#' @importFrom shinytoastr useToastr toastr_error toastr_success
+#' @importFrom shinytoastr useToastr toastr_error toastr_info toastr_success toastr_warning
 #' @importFrom shinylogs store_rds track_usage
-#' @importFrom shiny actionButton conditionalPanel fluidRow h4 isolate observe observeEvent onSessionEnded parseQueryString req tagList tags textInput updateTextInput
+#' @importFrom shiny actionButton conditionalPanel fluidRow h4 isolate observe observeEvent onSessionEnded parseQueryString req tagList tags textInput updateActionButton updateTextInput
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start