+++ NEWS.md
@@ -1,7 +1,14 @@
 # learndown 1.0.1
 
 - Shiny apps: do not track events when no login and fallback to temporary
-directory if `path =` is not writeable in `trackEvents()`.
+directory if `path =` is not writable in `trackEvents()`.
+
+- The number of opened sessions is tracked in a Shiny app and `stopApp()` is
+only invoked if there is no other opened session.
+
+- For Shiny applications, `trackEvents()` only tracks events in a MongoDB
+database for logged users. A message appears briefly (5sec) to indicate the
+status of the recording when the session is started.
 
 # learndown 1.0.0